Revised version of Labeling.
For separation of concerns, the functionality has been stripped down in
comparison to the old version: Labeling only represents an image whose
pixels are sets of labels of type T. Other capabilities such as getting
bounding boxes of particular labels etc. will be added in separate
classes that operate on the labeling.

Labeling<T> is a RandomAccessible and IterableInterval of LabelingType<T>.
LabelingType<T> is a set of labels of type T.
Currently all Labelings are based on LabelingMapping to map label sets
to integer values for efficient storage. ImgLabeling is an implementation
of Labeling, using a RandomAccessibleInterval of some integral type for
storage.

LabelingMapping maintains lookup tables that allow to perform addition
or removal of a label from a pixel directly on the integer values that
represent label sets. This avoids the costly steps of creating a
HashSet, adding the label, and interning the resulting set. This will
hopefully eliminate the need for explicit label set "interning" in user
code.

A dependency on trove4j has been added because we use TObjectIntMap for
the add-/remove-label lookup tables in LabelingMapping to avoid boxing/
unboxing of label set indices.