default method hierarchy for Mask, MaskInterval interfaces (and Real versions)
This is trying to as much as possible specify whether the result of some
operation is an interval (vs. unbounded) in the types.

For example (A and B) is bounded if at least one of A, B is bounded.
Therefore, in MaskInterval interface and() default method is overridden
  MaskInterval and( final Predicate<?> other )
because we know that one of the operators (this) is bounded. Therefore the
result must be bounded.

And so on for other operations and for real mask variants...