An iterable over a quasi Bresenham line.
It will iterate exactly once over all the integer locations on a
line between in proper order from the specified start to the
specified end points, included.

This implementation uses floating-point logic instead of the pure
integer logic of Bresenham line (Wikipedia) but the results are
quasi identical and the performance penalty small.